Fix version analyzer

+++ src/dotnet/Azure.ClientSdk.Analyzers/Azure.ClientSdk.Analyzers/ClientOptionsAnalyzer.cs
@@ -63,7 +63,7 @@ namespace Azure.ClientSdk.Analyzers
                     }
 
                     var maxVersion = serviceVersionEnum.GetMembers().Where(m => m.Kind == SymbolKind.Field).Max(m => ((IFieldSymbol)m).ConstantValue);
-                    if (!firstParam.HasExplicitDefaultValue || firstParam.ExplicitDefaultValue != maxVersion)
+                    if (!firstParam.HasExplicitDefaultValue || (int)firstParam.ExplicitDefaultValue != (int)maxVersion)
                     {
                         context.ReportDiagnostic(Diagnostic.Create(Descriptors.AZC0010, firstParam.Locations.First()), typeSymbol);
                     }
